Main entry point of each module has been modified.
